Implement Reassure stability check retry mechanism
