+++ .github/actions/javascript/reassureStabilityCheck/index.js
@@ -24919,14 +24919,14 @@ exports["default"] = _default;
 
 /***/ }),
 
-/***/ 2378:
+/***/ 4503:
 /***/ (function(__unused_webpack_module, exports, __nccwpck_require__) {
 
 "use strict";
 
 /*
  *  NOTE: After changes to the file it needs to be compiled using [`ncc`](https://github.com/vercel/ncc)
- *  Example: ncc build -t validateReassureOutput.ts -o index.js
+ *  Example: ncc build -t reassureStabilityCheck.ts -o index.js
  */
 var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
@@ -24960,89 +24960,22 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
 Object.defineProperty(exports, "__esModule", ({ value: true }));
 const core = __importStar(__nccwpck_require__(2186));
-const fs_1 = __importDefault(__nccwpck_require__(7147));
-function getInputOrEnv(name) {
-    try {
-        return core.getInput(name, { required: true });
-    }
-    catch (_a) {
-        const envProperty = process.env[name];
-        if (!envProperty) {
-            throw new Error(`'${name}' env property not defined.`);
-        }
-        return envProperty;
-    }
-}
+const child_process_1 = __nccwpck_require__(2081);
 function run() {
     return __awaiter(this, void 0, void 0, function* () {
         try {
-            const regressionOutput = JSON.parse(fs_1.default.readFileSync('.reassure/output.json', 'utf8'));
-            const allowedDurationDeviation = Number(getInputOrEnv('ALLOWED_DURATION_DEVIATION'));
-            const durationDeviationPercentage = Number(getInputOrEnv('ALLOWED_RELATIVE_DURATION_DEVIATION'));
-            const isValidatingStability = Boolean(getInputOrEnv('IS_VALIDATING_STABILITY'));
-            if (regressionOutput.significant === undefined || regressionOutput.significant.length === 0) {
-                console.log('No significant data available. Exiting...');
-                return true;
-            }
-            const outputs = [];
-            console.log(`Processing ${regressionOutput.significant.length} measurements...`);
-            for (let i = 0; i < regressionOutput.significant.length; i++) {
-                const index = i + 1;
-                const measurement = regressionOutput.significant[i];
-                const durationDeviation = measurement.durationDiff;
-                const relativeDurationDeviation = measurement.relativeDurationDiff;
-                const relativeDurationDeviationPercentage = relativeDurationDeviation * 100;
-                console.log(`Processing measurement ${index}: ${measurement.name}`);
-                const isMeasurementRelevant = Math.abs(durationDeviation) > allowedDurationDeviation;
-                if (!isMeasurementRelevant) {
-                    console.log(`Skipping measurement ${index} as it's not relevant.`);
-                    continue;
-                }
-                if (relativeDurationDeviationPercentage > durationDeviationPercentage) {
-                    outputs.push({
-                        name: measurement.name,
-                        description: `Duration deviation of ${durationDeviation.toFixed(2)} ms (${relativeDurationDeviationPercentage.toFixed(2)}%) exceeded the allowed range of ${allowedDurationDeviation.toFixed(2)} ms (${durationDeviationPercentage.toFixed(2)}%).`,
-                        relativeDurationDeviationPercentage,
-                        isDeviationExceeded: true,
-                    });
-                }
-                else {
-                    outputs.push({
-                        name: measurement.name,
-                        description: `Duration deviation of ${durationDeviation.toFixed(2)} ms (${relativeDurationDeviationPercentage.toFixed(2)}%) is within the allowed range of ${allowedDurationDeviation.toFixed(2)} ms (${durationDeviationPercentage.toFixed(2)}%).`,
-                        relativeDurationDeviationPercentage,
-                        isDeviationExceeded: false,
-                    });
-                }
-            }
-            if (outputs.length === 0) {
-                console.log('No relevant measurements. Exiting...');
-                return true;
-            }
-            console.log('\nSummary:');
-            outputs.sort((a, b) => b.relativeDurationDeviationPercentage - a.relativeDurationDeviationPercentage);
-            outputs.forEach((output) => {
-                console.log(`${output.isDeviationExceeded ? '🔴' : '🟢'} ${output.name} > ${output.description}`);
-            });
-            const shouldFailWorkflow = outputs.some((output) => output.isDeviationExceeded);
-            if (shouldFailWorkflow) {
-                if (isValidatingStability) {
-                    core.setFailed(`🔴 Duration deviation exceeded the allowed ranges in one or more measurements during the stability checks.`);
-                }
-                else {
-                    core.setFailed(`🔴 Duration deviation exceeded the allowed ranges in one or more measurements.`);
-                }
-            }
+            console.log('Running Reassure stability check...');
+            (0, child_process_1.execSync)('npx reassure check-stability --verbose', { stdio: 'inherit' });
+            console.log('Validating Reassure stability results...');
+            (0, child_process_1.execSync)('node .github/actions/javascript/validateReassureOutput', { stdio: 'inherit' });
             return true;
         }
         catch (error) {
             console.log('error: ', error);
             core.setFailed(error.message);
+            process.exit(1);
         }
     });
 }
@@ -25073,6 +25006,14 @@ module.exports = require("async_hooks");
 
 "use strict";
 module.exports = require("buffer");
+
+/***/ }),
+
+/***/ 2081:
+/***/ ((module) => {
+
+"use strict";
+module.exports = require("child_process");
 
 /***/ }),
 
@@ -26943,7 +26884,7 @@ module.exports = parseParams
 /******/ 	// startup
 /******/ 	// Load entry module and return exports
 /******/ 	// This entry module is referenced by other modules so it can't be inlined
-/******/ 	var __webpack_exports__ = __nccwpck_require__(2378);
+/******/ 	var __webpack_exports__ = __nccwpck_require__(4503);
 /******/ 	module.exports = __webpack_exports__;
 /******/ 	
 /******/ })()
